requirements.txt
Flask==3.0.3
py_eureka_client==0.11.10
config==0.5.1
pdfkit==1.0.0
qrcode==7.4.2
PyPDF2==2.12.1
pillow==10.3.0
python-dotenv==1.0.1
gunicorn==22.0.0